requirements.txt
click==7.1.2
colorama==0.4.4
pydantic==1.10.4
commlib-py>=0.10.6
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
textX==3.1.1
rich==13.4.2
redis==3.5.3
python-multipart==0.0.6
